Document Important 3D Hardware Functions (#28)
This PR tries to partially solve issue #6 by documenting the overall usage of the 3D hardware. Giving brief explanations of the Geometry and Rendering Engine, as well as defining each available graphics function and it's parameter types/formats. 

This PR does not document certain intricacies of non-standard behaviour, and may or may not need gramatical refinements and touches in the future. It also does not provide details on the specs of the 3D hardware, or how graphics commands result in the final Vertex/Polygons RAM buffers.